Preventing Users from Self-Deletion

+++ server/controllers/userController.js
@@ -6,6 +6,7 @@ import bcrypt from "bcrypt"
 import Session from "../models/sessionModel.js";
 import OTP from "../models/otpModel.js";
 import File from "../models/fileModel.js";
+import { error } from "node:console";
 
 
 
@@ -91,7 +92,7 @@ export const login = async (req, res, next) => {
       error: "Your account has been deleted. Contact App Admin"
     });
   }
-  
+
   const isPasswordValid = await bcrypt.compare(password, user.password);
   // OR
   // const isPasswordValid = await user.comparePassword(password);
@@ -172,10 +173,10 @@ export const logoutAll = async(req, res) => {
 
 export const deleteUser = async(req, res , next) => {
   const {userId} = req.params
+  if(userId === req.user._id.toString()){
+    return res.status(403).json({error : "Cannot delete yourself"})
+  }
   try {
-    // await User.deleteOne({_id : userId})
-    // await File.deleteMany({userId})
-    // await Directory.deleteMany({userId})
     await Session.deleteMany({userId : req.params.userId})
     await User.findByIdAndUpdate(userId , {deleted : true})
     res.status(204).end();
